`SyncRawMixin` improved code readability
