+++ port_ocean/core/integrations/mixins/__init__.py
@@ -1,7 +1,7 @@
 from .events import EventsMixin
 from .handler import HandlerMixin
-from .sync import SyncRawMixin, SyncMixin
-
+from .sync import SyncMixin
+from .sync_raw import SyncRawMixin
 
 __all__ = [
     "EventsMixin",
